+++ ecommerce-back/src/main/java/com/ecommerce/dto/PurchaseOrderDetail/PurchaseOrderDetailDTO.java
@@ -5,14 +5,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class PurchaseOrderDetailDTO {
     private Integer quantity;
-    private Double unitPrice;
-    private Double totalPrice;
+    private BigDecimal unitPrice;
+    private BigDecimal totalPrice;
     private Long productVariantId;
     private Long discountId;
 }
